bugfix: Fix Column Index Backtracking
Because both `rows` and `cols` in `SolveCell()` are of type `size_t`,
decrementing past 0 resulted in an incorrect positioning for backtracking. As a
solution, once `col` reaches zero, it will become the last position in a row
(i.e. COLS - 1). Otherwise, it will decrement as normal.

Additionally, the bounds constraint placed while decrementing both `row` and
`col` was removed. Assuming the initial cells of the board follow Sudoku rules,
`row` will never loop back from the last row to the first, as intended by the
bound constraint. Likewise, the new conditional for decrementing `col` properly
performs a loop back from the last column of a given row to the last column of
its previous row.