test: Commit Test Suite
Since there is no sole correct answer for most sudokus, there is no way to
perform a diff to verify the correctness of the algorithm. For this reason,
`IsValid()` is called on every cell after a given board has been solved with
`SolveSudoku()`.

Parallely, the bash scripts iterates over the test puzzles in `puzzles/` and
runs them through `test_sudoku.c`, generating a report of passed vs failed test
cases.