bugfix: Handle Loading Boards With Invalid Dimensions
While `LoadBoards()` only reads up to `kRows` rows, the previous implementation
did not account for rows with more than `kCols` characters.

The new implementation skips characters once more than `kCols` characters have
been read in a given row. To test this, an invalid test case was added. It is
worth noting that this test case fails due to an error unrelated to this commit.
This bug will be patched in the next commit.