feat: Implement Input Board Validation
`IsBoardValid()` calls `IsValid()` on every cell to determine the correctness of
the input board. This check enhances robustness of the program by accounting for
all user inputs, either valid or invalid.

Readability improved by coloring the cells that violate Sudoku rules. The
approach for this was to add a new attribute to the `SudokuCell` struct,
a boolean `invalid`. `IsBoardValid()` changes the value of a given cell to
`false` when calling `IsValid()` on that cell returns `false`. This way,
formatting of valid and invalid cells by checking this attribute.

Cheers.
Probably the last commit...